Move secret rotation function into own package
